+++ manifest.json
@@ -18,8 +18,8 @@
         "https://chatgpt.com/*"
       ],
       "js": [
-        "ctrl-enter-utils.js",
-        "ctrl_enter_chatgpt.js"
+        "content/ctrl-enter-utils.js",
+        "content/ctrl-enter-chatgpt.js"
       ]
     },
     {
@@ -33,8 +33,8 @@
         "https://copilot.microsoft.com/*"
       ],
       "js": [
-        "ctrl-enter-utils.js",
-        "ctrl_enter_textarea.js"
+        "content/ctrl-enter-utils.js",
+        "content/ctrl-enter-textarea.js"
       ],
       "run_at": "document_start"
     },
@@ -49,8 +49,8 @@
         "https://grok.com/*"
       ],
       "js": [
-        "ctrl-enter-utils.js",
-        "ctrl_enter_custom_inputs.js"
+        "content/ctrl-enter-utils.js",
+        "content/ctrl-enter-custom-inputs.js"
       ],
       "run_at": "document_start"
     }
@@ -77,12 +77,12 @@
     "https://grok.com/*"
   ],
   "action": {
-    "default_popup": "popup.html",
+    "default_popup": "popup/popup.html",
     "default_icon": {
       "16": "icon/icon16.png",
       "48": "icon/icon48.png",
       "128": "icon/icon128.png"
     }
   },
-  "options_page": "options.html"
+  "options_page": "options/options.html"
 }
