+++ options/options.html
@@ -4,7 +4,7 @@
 <head>
   <meta charset="UTF-8">
   <title>ChatGPT Ctrl+Enter Sender Settings</title>
-  <link rel="stylesheet" href="options.css">
+  <link rel="stylesheet" href="./options.css">
 </head>
 
 <body>
@@ -15,7 +15,7 @@
 
   <button id="saveButton">Save</button>
 
-  <script type="module" src="options.js"></script>
+  <script type="module" src="./options.js"></script>
 </body>
 
 </html>
