+++ popup/popup.html
@@ -4,7 +4,7 @@
 <head>
   <meta charset="UTF-8">
   <title>ChatGPT Ctrl+Enter Sender Options</title>
-  <link rel="stylesheet" href="popup.css">
+  <link rel="stylesheet" href="./popup.css">
 </head>
 
 <body>
@@ -19,13 +19,13 @@
       </div>
     </div>
 
-    <a href="options.html" id="settingsLink" target="_blank" class="settings-link">
+    <a href="../options/options.html" id="settingsLink" target="_blank" class="settings-link">
       All Site Settings
       <span class="external-icon">↗</span>
     </a>
   </div>
 
-  <script type="module" src="popup.js"></script>
+  <script type="module" src="./popup.js"></script>
 </body>
 
 </html>
